rename components directory
